20-eatopia-agents/data_pac_ia/requirements.txt
# Dependências para o Data Pac IA
adk>=0.1.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Para a API mock (opcional, apenas para testes)
flask>=3.0.0 